Feat: add shopId and isDeleted fields to Product model; update related services and routes

+++ src/app/modules/Product/product.service.ts
@@ -16,18 +16,25 @@ const createProductIntoDb = async (vendorId: string, cloudinaryResult: any, data
       vendor: true
     }
   });
-  if (!vendor || !vendor.vendor) throw new ApiError(StatusCodes.NOT_FOUND, "Vendor not found!");
+  if (!vendor || !vendor.vendor) throw new ApiError(StatusCodes.NOT_FOUND, "User not found!");
+  const isShopExist = await prisma.shop.findUnique({
+    where: {
+      id: data.shopId,
+      isDeleted: false
+    }
+  });
+  if (!isShopExist) throw new ApiError(StatusCodes.NOT_FOUND, "Shop not found!");
   const category = await prisma.category.findUnique({
     where: {
       id: data.categoryId,
     }
   });
   if (!category) throw new ApiError(StatusCodes.NOT_FOUND, "Category not found!");
-  if (vendorId) {
-    data.vendorId = vendorId
-  }
   if (cloudinaryResult && cloudinaryResult.secure_url) {
     data.image = cloudinaryResult.secure_url;
+  }
+  if (vendorId) {
+    data.vendorId = vendorId;
   }
   const result = await prisma.product.create({
     data
@@ -37,7 +44,15 @@ const createProductIntoDb = async (vendorId: string, cloudinaryResult: any, data
 
 // get all products
 const getAllProductsFromDb = async () => {
-  const result = await prisma.product.findMany({});
+  const result = await prisma.product.findMany({
+    where: {
+      isDeleted: false
+    },
+    include: {
+      category: true,
+      shop: true,
+    }
+  });
   return result;
 }
 
@@ -45,23 +60,37 @@ const getAllProductsFromDb = async () => {
 const getSingleProductFromDb = async (id: string) => {
   const result = await prisma.product.findUniqueOrThrow({
     where: {
-      id
+      id,
+      isDeleted: false
+    },
+    include: {
+      category: true,
+      shop: true,
     }
   });
   return result;
 };
 
-// get vendor all products
-const getVendorAllProductsFromDb = async (vendorId: string) => {
-  await prisma.vendor.findUniqueOrThrow({
+// get shop all products
+const getShopAllProductsFromDb = async (shopId: string) => {
+  const shop = await prisma.shop.findUnique({
     where: {
-      userId: vendorId,
+      id: shopId,
       isDeleted: false
+    },
+    include: {
+      vendor: true
     }
-  })
+  });
+  if (!shop) throw new ApiError(StatusCodes.NOT_FOUND, "Shop not found!");
   const result = await prisma.product.findMany({
     where: {
-      vendorId
+      shopId,
+      isDeleted: false
+    },
+    include: {
+      category: true,
+      shop: true,
     }
   });
   return result;
@@ -69,24 +98,24 @@ const getVendorAllProductsFromDb = async (vendorId: string) => {
 
 // update product
 const updateProductIntoDb = async (vendorId: string, cloudinaryResult: any, productId: string, data: Partial<IProduct>) => {
-  const currentVendor = await prisma.vendor.findUniqueOrThrow({
+  const currentVendor = await prisma.user.findUniqueOrThrow({
     where: {
-      userId: vendorId,
+      id: vendorId,
+      role: UserRole.VENDOR,
       isDeleted: false
-    },
-    select: {
-      userId: true
     }
   });
-  const currentProduct = await prisma.product.findUniqueOrThrow({
+  const currentProduct = await prisma.product.findUnique({
     where: {
-      id: productId
+      id: productId,
+      isDeleted: false
     },
     select: {
       vendorId: true
     }
-  })
-  if (currentProduct?.vendorId !== currentVendor?.userId) throw new ApiError(StatusCodes.UNAUTHORIZED, "You are not authorized to update this product!");
+  });
+  if (!currentProduct) throw new ApiError(StatusCodes.NOT_FOUND, "Product not found!");
+  if (currentProduct.vendorId !== currentVendor.id) throw new ApiError(StatusCodes.UNAUTHORIZED, "You are not authorized!");
   if (data.categoryId) {
     await prisma.category.findUniqueOrThrow({
       where: {
@@ -95,7 +124,7 @@ const updateProductIntoDb = async (vendorId: string, cloudinaryResult: any, prod
     });
   }
   if (vendorId) {
-    data.vendorId = vendorId
+    data.vendorId = vendorId;
   }
   if (cloudinaryResult && cloudinaryResult.secure_url) {
     data.image = cloudinaryResult.secure_url;
@@ -111,21 +140,25 @@ const updateProductIntoDb = async (vendorId: string, cloudinaryResult: any, prod
 
 // delete product
 const deleteProductFromDb = async (productId: string, userId: string) => {
-  const currentUser = await prisma.user.findUniqueOrThrow({
+  const currentUser = await prisma.user.findUnique({
     where: {
-      id: userId
+      id: userId,
+      isDeleted: false
     }
-  })
+  });
+  if (!currentUser) throw new ApiError(StatusCodes.NOT_FOUND, "User not found!");
   const currentProduct = await prisma.product.findUniqueOrThrow({
     where: {
-      id: productId
+      id: productId,
+      isDeleted: false
     },
     select: {
       vendorId: true
     }
-  })
+  });
+  if (!currentProduct) throw new ApiError(StatusCodes.NOT_FOUND, "Product not found!");
   const isAdmin = currentUser.role === UserRole.ADMIN || currentUser.role === UserRole.SUPER_ADMIN;
-  const isAuthorized = currentProduct?.vendorId === userId || isAdmin
+  const isAuthorized = currentProduct?.vendorId === userId || isAdmin;
   if (!isAuthorized) throw new ApiError(StatusCodes.UNAUTHORIZED, "You are not authorized!");
   const result = await prisma.product.delete({
     where: {
@@ -139,7 +172,7 @@ export const ProductServices = {
   createProductIntoDb,
   getAllProductsFromDb,
   getSingleProductFromDb,
-  getVendorAllProductsFromDb,
+  getShopAllProductsFromDb,
   updateProductIntoDb,
   deleteProductFromDb
 }
